requirements.txt
openai-whisper
google-generativeai 
streamlit 
python-dotenv
ffmpeg